+++ ai/integrations/llama_index/tests/test_llama_index_toolkit_oss.py
@@ -123,7 +123,7 @@ def generate_mock_execution_result(return_value: str = "result") -> FunctionExec
 
 def test_toolkit_creation_errors_no_client(monkeypatch):
     monkeypatch.setattr(
-        "unitycatalog.ai.core.utils.client_utils._is_databricks_client_available", lambda: False
+        "unitycatalog.ai.core.base._is_databricks_client_available", lambda: False
     )
 
     with pytest.raises(
